requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
google-generativeai
assemblyai

fastapi==0.95.2
uvicorn[standard]==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
google-generative-ai==0.5.1   # if using Gemini via official client
assemblyai==1.2.1
serpapi==3.0.1
